pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agent"
dynamic = ["version"]
description = 'A project managed by Hatch'
readme = "docs/README.md"
requires-python = ">=3.8"
license = "MIT"
keywords = []
authors = [
  { name = "RodePachi", email = "pachirode@gmail.com" },
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: Implementation :: CPython",
  "Programming Language :: Python :: Implementation :: PyPy",
]
dependencies = [
  "requests",
  "tavily-python",
  "openai",
  "python-dotenv",
  "google-search-results",
  "pyparsing",
]

[project.urls]
Documentation = "https://github.com/unknown/agent#readme"
Issues = "https://github.com/unknown/agent/issues"
Source = "https://github.com/unknown/agent"

[tool.hatch.version]
path = "src/agent/__about__.py"

[tool.hatch.envs.default]
type = "virtual"
dependencies = [
  "pytest",
  "pytest-cov",
]
env-vars = { PYTHONPATH = "src" }
[tool.hatch.envs.default.scripts]
test = "pytest src/tests"
test-cov = "pytest --cov-report=term-missing --cov-config=pyproject.toml --cov=src/agent --cov=tests {args:tests}"
typecheck = "mypy --install-types --non-interactive {args:src/agent tests}"

[tool.hatch.build.targets.sdist]
exclude = [
  "/.github",
]

[tool.hatch.build.targets.wheel]
packages = ["src/agent"]
